# Improve performance of `DataGroup.shape`